mysite/requirements.txt
astroid==2.3.0
beautifulsoup4==4.8.0
Django==2.2.5
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap3==11.1.0
django-bootstrap4==1.0.1
django-crispy-forms==1.7.2
django-easy-pdf==0.1.1
django-mathfilters==0.4.0
html5lib==1.0.1
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
Pillow==6.1.0
pkg-resources==0.0.0
pylint==2.4.0
PyPDF2==1.26.0
pytz==2019.2
reportlab==3.5.26
six==1.12.0
soupsieve==1.9.3
sqlparse==0.3.0
typed-ast==1.2.0
webencodings==0.5.1
wrapt==1.11.2
xhtml2pdf==0.2.3
